- refactor the code & make sure only one OPA query per request - use Akka http client for http request - Avoid passing JWT token via Query case class - Merge MatchNoneQuery - When no publishingState filter in place, query OPA with wildcard operation URI
