+++ magda-search-api/src/main/scala/au/csiro/data61/magda/search/SearchQueryer.scala
@@ -7,8 +7,8 @@ import au.csiro.data61.magda.model.misc._
 import scala.concurrent.Future
 
 trait SearchQueryer {
-  def search(query: Query, start: Long, limit: Int, facetSize: Int): Future[SearchResult]
-  def searchFacets(facetType: FacetType, facetQuery: Option[String], generalQuery: Query, start: Int, limit: Int): Future[FacetSearchResult]
+  def search(jwtToken: Option[String], query: Query, start: Long, limit: Int, facetSize: Int): Future[SearchResult]
+  def searchFacets(jwtToken: Option[String],facetType: FacetType, facetQuery: Option[String], generalQuery: Query, start: Int, limit: Int): Future[FacetSearchResult]
   def searchRegions(query: Option[String], start: Long, limit: Int): Future[RegionSearchResult]
   def searchOrganisations(query: Option[String], start: Int, limit: Int): Future[OrganisationsSearchResult]
 }
